not fix test

+++ etl-service/src/main/java/com/sitefilm/etl/configuration/client/CoreTmdbClient.java
@@ -53,7 +53,7 @@ public class CoreTmdbClient {
                 .body(PersonsInMovieResponseDto.class);
     }
 
-    public PersonDetailsResponseDto loadPersonDetails(Long personId) {
+    public PersonDetailsResponseDto loadPersonDetails(Integer personId) {
         System.out.println("Запрос на персон детали");
         return restClient.get()
                 .uri(uriBuilder ->
